Change how string is rendered
It now renderes the string using a single `state` called `ind`, which is
set initially when the `devString` is changed/received. And the other
`useEffect` hook has `ind` as the dependency, which sets the
`currentString`. The `ind` goes from `+length_of_devString` to
`-length_of_devString`. When it reaches the `-ve` value,
`updatePosition` is called, which now starts rendering the next dev
role.